backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
python-dotenv==1.0.1
openai>=1.52.0
qdrant-client==1.12.1
python-multipart==0.0.20
PyYAML==6.0.1
markdown==3.5.2
beautifulsoup4==4.12.3
unstructured>=0.15.0
tiktoken==0.7.0
numpy==1.26.3
pandas==2.1.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic-settings==2.6.1
betterproto==2.0.0b6
psycopg2-binary==2.9.9
neon-python==0.3.0
httpx==0.27.0
requests==2.32.3